modelディレクトリを削除
