+++ lib/pages/my_home_page.dart
@@ -2,9 +2,55 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_tts/flutter_tts.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:mobile_translation_app/models/language.dart';
 import 'package:mobile_translation_app/views/text_field_item_view.dart';
 import 'package:translator/translator.dart';
+
+enum Language {
+  english,
+  japanese,
+}
+
+extension LanguageExtension on Language {
+  // 画面に表示するテキスト
+  String get viewText {
+    switch (this) {
+      case Language.english:
+        return 'English';
+      case Language.japanese:
+        return '日本語';
+    }
+  }
+
+  // translatorで翻訳言語を指定するラベル
+  String get translatorLabel {
+    switch (this) {
+      case Language.english:
+        return 'en';
+      case Language.japanese:
+        return 'ja';
+    }
+  }
+
+  // flutter_ttsで言語を指定するラベル
+  String get speakingLabel {
+    switch (this) {
+      case Language.english:
+        return 'en-US';
+      case Language.japanese:
+        return 'ja-JP';
+    }
+  }
+
+  // 言語によってキーボードの種類を変える
+  TextInputType get textInputType {
+    switch (this) {
+      case Language.japanese:
+        return TextInputType.text;
+      default:
+        return TextInputType.emailAddress;
+    }
+  }
+}
 
 class MyHomePage extends StatefulWidget {
   const MyHomePage({Key? key}) : super(key: key);
@@ -31,18 +77,23 @@ class _MyHomePageState extends State<MyHomePage> {
     setState(() {});
   }
 
+  // 翻訳を行うメソッド
+  // 引数isTopFieldで上側のTextFieldか下側のTextFieldか判別
   void _translate(String inputText, {required bool isTopField}) async {
     if (inputText.isEmpty) {
       return;
     }
     final translator = GoogleTranslator();
+    // 翻訳後の言語を指定
     final String to = isTopField
         ? _bottomLanguage.translatorLabel
         : _topLanguage.translatorLabel;
+    // 入力されたテキストinputTextを言語toに翻訳
     final translation = await translator.translate(
       inputText,
       to: to,
     );
+    // 翻訳結果をUIに反映
     if (isTopField) {
       _bottomController.text = translation.text;
     } else {
